# src/main/resources/static/swagger.yaml
openapi: 3.0.3
info:
  title: Myth Cloth API
  version: 1.0.0
  description: REST API documentation generated from the current controllers and DTOs.
servers:
  - url: /api/v1
    description: Relative to current host
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /catalogs/{catalogType}:
    parameters:
      - $ref: '#/components/parameters/CatalogTypePathParam'
    get:
      tags: [Catalogs]
      summary: List catalog entries by type
      operationId: retrieveCatalogs
      responses:
        '200':
          description: List of catalog entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CatalogResp'
        default:
          $ref: '#/components/responses/ProblemResponse'
    post:
      tags: [Catalogs]
      summary: Create catalog entry
      operationId: createCatalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatalogReq'
      responses:
        '201':
          description: Catalog entry created
          headers:
            Location:
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResp'
        default:
          $ref: '#/components/responses/ProblemResponse'
  /catalogs/{catalogType}/{id}:
    parameters:
      - $ref: '#/components/parameters/CatalogTypePathParam'
      - $ref: '#/components/parameters/IdPathParam'
    get:
      tags: [Catalogs]
      summary: Get catalog entry by id
      operationId: retrieveCatalog
      responses:
        '200':
          description: Catalog entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResp'
        default:
          $ref: '#/components/responses/ProblemResponse'
    put:
      tags: [Catalogs]
      summary: Update catalog entry
      operationId: updateCatalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatalogReq'
      responses:
        '200':
          description: Updated catalog entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResp'
        default:
          $ref: '#/components/responses/ProblemResponse'
    delete:
      tags: [Catalogs]
      summary: Delete catalog entry
      operationId: removeCatalog
      responses:
        '204':
          description: Catalog entry deleted
        default:
          $ref: '#/components/responses/ProblemResponse'
  /distributors:
    get:
      tags: [Distributors]
      summary: List distributors
      operationId: retrieveDistributors
      responses:
        '200':
          description: List of distributors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DistributorResp'
        default:
          $ref: '#/components/responses/ProblemResponse'
    post:
      tags: [Distributors]
      summary: Create distributor
      operationId: createDistributor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DistributorReq'
      responses:
        '201':
          description: Distributor created
          headers:
            Location:
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorResp'
        default:
          $ref: '#/components/responses/ProblemResponse'
  /distributors/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    get:
      tags: [Distributors]
      summary: Get distributor by id
      operationId: retrieveDistributor
      responses:
        '200':
          description: Distributor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorResp'
        default:
          $ref: '#/components/responses/ProblemResponse'
    put:
      tags: [Distributors]
      summary: Update distributor
      operationId: updateDistributor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DistributorReq'
      responses:
        '200':
          description: Updated distributor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorResp'
        default:
          $ref: '#/components/responses/ProblemResponse'
    delete:
      tags: [Distributors]
      summary: Delete distributor
      operationId: removeDistributor
      responses:
        '204':
          description: Distributor deleted
        default:
          $ref: '#/components/responses/ProblemResponse'
...
  /collections/{collectionId}/figurines/{figurineId}:
    parameters:
      - $ref: '#/components/parameters/CollectionIdPathParam'
      - $ref: '#/components/parameters/FigurineIdPathParam'
    put:
      tags: [Collections]
      summary: Add figurine to collection
      operationId: addFigurineToCollection
      responses:
        '200':
          description: Figurine added to collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FigurineResp'
        default:
          $ref: '#/components/responses/ProblemResponse'
components:
  parameters:
    IdPathParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    FigurineIdPathParam:
      in: path
      name: figurineId
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    CatalogTypePathParam:
      in: path
      name: catalogType
      required: true
      schema:
        type: string
        enum: [distributions, groups, lineups, series]
...
    CollectionIdPathParam:
      in: path
      name: collectionId
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
    ProblemResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        timestamp:
          type: string
          format: date-time
        errors:
          type: object
          additionalProperties:
            type: string
        error:
          type: string
...
    AnniversaryResp:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        year:
          type: integer
    CatalogReq:
      type: object
      required: [description]
      properties:
        description:
          type: string
          maxLength: 100
    CatalogResp:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
    DistributorReq:
      type: object
      required: [name, countryCode]
      properties:
        name:
          type: string
          enum: [BANDAI, DAM, DTM, BANDAI_CHINA, DS_DISTRIBUTIONS, BLUE_FIN]
        countryCode:
          type: string
          enum: [JP, MX, ES, US, CN]
        website:
          type: string
    DistributorResp:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        countryCode:
          type: string
          enum: [JP, MX, ES, US, CN]
        website:
          type: string
...
    FigurineDistributorResp:
      type: object
      properties:
        distributor:
          $ref: '#/components/schemas/DistributorResp'
        currency:
          type: string
          enum: [JPY, MXN, EUR, USD, CNY, CAD]
        price:
          type: number
          format: double
        priceWithTax:
          type: number
          format: double
        announcedAt:
          type: string
          format: date
        preorderOpensAt:
          type: string
          format: date
        releaseDate:
          type: string
          format: date
        releaseDateConfirmed:
          type: boolean
...
    FigurineEventResp:
      type: object
      properties:
        id:
          type: integer
          format: int64
        date:
          type: string
          format: date
        type:
          type: string
          enum: [ANNOUNCEMENT, PREORDER_OPEN, PREORDER_CLOSE, RELEASE, RESTOCK, LOCAL_CONFIRMATION, LOCAL_RELEASE]
        region:
          type: string
          enum: [JP, MX, ES, US, CN]
        description:
          type: string
...
    FigurineResp:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        displayableName:
          type: string
        distributors:
          type: array
          items:
            $ref: '#/components/schemas/FigurineDistributorResp'
        tamashiiUrl:
          type: string
        distribution:
          $ref: '#/components/schemas/CatalogResp'
        lineUp:
          $ref: '#/components/schemas/CatalogResp'
        series:
          $ref: '#/components/schemas/CatalogResp'
        group:
          $ref: '#/components/schemas/CatalogResp'
        anniversary:
          $ref: '#/components/schemas/AnniversaryResp'
        isMetalBody:
          type: boolean
        isOriginalColorEdition:
          type: boolean
        isRevival:
          type: boolean
        isPlainCloth:
          type: boolean
        isBattleDamaged:
          type: boolean
        isGoldenArmor:
          type: boolean
        isGold24kEdition:
          type: boolean
        isMangaVersion:
          type: boolean
        isMultiPack:
          type: boolean
        isArticulable:
          type: boolean
        notes:
          type: string
        officialImageUrls:
          type: array
          items:
            type: string
        unofficialImageUrls:
          type: array
          items:
            type: string
        events:
          type: array
          items:
            $ref: '#/components/schemas/FigurineEventResp'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...